Add tos inherit to the various ipv6 tunneling methods
Without this tos and ecn bits are not copied from the inner header
to the outer. Having those bits present is helpful for those doing
shaping based on the tos bits, and using ECN in their tunnels.

(it's also a nifty check to see if your provider is messing with
 your tos bits)